Uniformly use map-reduce kernels.
Previously, we had to distinguish two kinds of kernels: Those containing
reduction statements, and those without.  This introduced some amount of
code duplication and confusion.  So now we introduce an artificial dummy
reduction in all kernels that would otherwise not have a reduction.

In doing so, we can finally incorporate the reduction range into the
iteration space of the kernel.  By convention, the first range of each
kernel is the reduction range.  Thanks to this uniform treatment, we can
now perform IR normalization on reduction kernels, too.

The only thing we have to make sure now is that when we rotate the kernel
iteration space, we must not change the position of the reduction range.