Introduce a new, vastly improved type inference engine.
The new type inference code can also handle EQL types.  It borrows a lot of
ideas from the previous petalisp.specialization code in that it performs
type inference as a byproduct of rewriting complicated functions into
several simpler ones.  We already use this automatic specialization in the
application constructor, such that applications of known functions are
turned into multiple nested calls to simpler functions.  This extra
information should be useful for some backends.

The next steps are:
- Perform constant folding whenever all arguments are EQL types.
- Also use specializations for reductions.