Rename coerce-to-lazy-array to lazy-array.
The former is more precise, but the function is used so frequently that a
shorter form seems appropriate.  In doing so, we also follow the convention
of functions like cl:string and cl:pathname, where a function of the same
name as the type coerces an object designator to the actual object.