Write a version of β that doesn't use reductions.
However, we don't use this new version until it has been thoroughly
tested and shown to be similarly fast as the old version.

The great news is that as soon as the new version is deemed viable, we can
eliminate reductions from the Petalisp core.  This should simplify the code
and all backends a lot.