Use lazy array layouts to improve IR conversion.
The IR still consists of buffers and layouts, but now there can be more
than one buffer per lazy array.  Usually, one buffer is created per kernel
and load instruction, except that loads from a single kernel and with
similar transformations are grouped together and share a single buffer.

The resulting IR contains many more buffers and auxiliary copy kernels, but
the improved granularity also allows for more optimizations.  By tracking
producers and users in more detail, it is often possible to identify new
dead code.

The main reason for introducing this change is that it greatly simplifies
partitioning a kernel for parallel execution.