Apply IR optimizations in the ideal order.
When finalizing IR layouts, consecutive copy kernels are automatically
merged.  However, this optimization fails to merge copy kernels in case a
layout is finalized before the layouts of its inputs have been finalized.
The solution is to sort all layouts by their depth in the original data
flow graph, and to process them starting from the ones with the lowest
depth.

Conveniently, this sorted array of layouts also allows us to remove buffers
that are referenced zero times starting from the ones with the highest
depth.