MAJOR INCOMPATIBLE CHANGE: New range syntax.
This change has already been discussed with members of the community.  It
changes the syntax for ranges and shapes such that the upper bound is
always exclusive and such that the step size is now the third argument.
Also, if only a single integer is supplied it denotes a range starting from
zero up to, but excluding, that integer.

I chose the old notation because I didn't want to pressure people
syntactically into starting their arrays with zero.  But there are also
very good reasons for the new notation.  For one, it makes Petalisp easier
to use.  The new notation behaves almost exactly like the range notation in
Python and most other programming languages.  The other good reason is that
the new syntax increases productivity when using both Lisp and Petalisp,
since Lisp arrays always start with zero.

This change is more invasive than one would think, because by having an
exclusive upper bound, it is now possible to have empty ranges and empty
shapes.  But having these new data structures is a good thing anyway.  It
allowed me to remove several kludges where I was using NIL to denote the
empty range or the empty shape.

I should also note that this change has zero impact on the expressiveness
of Petalisp.  It is (apart from the empty ranges and shapes) a purely
cosmetic change.  Apologies to everyone whose programs I just crippled.

Oh, and a short upgrade guide for everyone with a broken program:

Simply replace (~ A B C) by (~ A (1+ C) B).