MAJOR INCOMPATIBLE CHANGE: Change semantics of range-end.
This commit is strongly related to the previous one.  The function
RANGE-END will now return the element one larger than the last one, to be
more consistent with how Common Lisp usually handles the END argument of
intervals.  For those who want the old behavior of returning the last
element, there is now a new function called RANGE-LAST.