Implement a new IR conversion algorithm.
This commit doesn't affect the IR itself, only how it is generated.  The
new IR conversion algorithm is much shorter, more readable, and it is
better at handling a variety of corner cases.  Also, conversion is now a
single pass over the supplied lazy arrays.

This commit also changes the handling of lazy multiple value arrays and the
reference implementation.  Also, it removes the ASSOCF utility in favor of
alexandria's ASSOC-VALUE.