Miscellaneous IR conversion improvements.
There is now a single instruction both for calls with one value and calls
with multiple values.  This simplifies the task of writing a backend.

This commit fixes a subtle bug where a dendrite's cons cell won't contain a
suitable instruction after growing further.  These cells are now updated in
a separate IR conversion pass.

The IR checker will now check instructions more carefully.

The ntype slot of a lazy multiple value map instruction will now contain a
list of ntypes - one for each of its outputs.  This information might be
used for some future optimizations.

Validity of stems and dendrites is now stored in the stem, not in each
dendrite.  This avoids some special code for copying dendrites and also
quite some consing.