Change blueprints and write a new blueprint compiler.
The new blueprints do not necessarily hard-code the offsets and strides of
a transformation.  As a result, there is no need to generate different
kernels in case they only differ in some offsets of some transformation.

The new blueprint grammar also required a change of the blueprint compiler.
Since the old one was not exactly a beauty, I decided to rewrite it from
scratch.  The new blueprint compiler should be much easier to grasp and to
maintain.

The new blueprint compiler has been merged into the petalisp.ir package, so
the old native-compiler system has become obsolete.