Write a new kernel interpreter.
The new interpreter exploits the fact that kernels do now have an
instruction vector that is sorted in order of the instruction dependencies.
The new interpreter is faster and conses less.

Also, the new kernel interpreter is now part of the petalisp.ir package, so
that it can be used by multiple backends.