Split the behavior of lisp-datum-from-immediate.
From now on, there is a generic function called array-from-immediate for
converting immediates to Lisp arrays.  The function
lisp-datum-from-immediate is not generic anymore - all it does is calling
array-from-immediate and converting the result to a scalar if it has rank
zero.

The motivation for this change is that we expect backends to provide their
own kinds of immediates, and methods from turning these immediates into
standard arrays.  These custom methods should not be bothered with
unpacking of rank zero arrays.