New optimization: Combine multiple steps of a reduction.
This new optimization reduces the memory footprint of reductions, by
incorporating multiple reduction steps into one kernel.  A keyword argument
supplied to ir-from-lazy-arrays determines how much a kernel is allowed to
grow when applying this optimization.

This commit also improves the IR checker, and introduces a function for
accessing the size of a buffer.