MAJOR INCOMPATIBLE CHANGE: Introduce a new naming convention.
The new naming convention uniformly uses a LAZY- prefix for functions that
return lazy arrays.  So SLICE is now called LAZY-SLICE, FLATTEN is now
called LAZY-FLATTEN, RESHAPE is now called LAZY-RESHAPE, and so on.

Also, this commit gets rid of Greek letters in the API.  The original
motivation for using Greek letters was to use 'α' to introduce parallelism
of the first kind, and 'β' to introduce parallelism of the second kind.
But since we recently managed to implement 'β' in terms of 'α', this naming
convention doesn't make sense anymore.  Also, the Greek letters have been a
frequent source of frustration for new users without fancy input methods.

Here is how α, β, and τ are written from now on:

The function 'α' is now called 'lazy'.

The function 'α+' is now called 'lazy-multiple-value'.

The function 'β' is now called 'lazy-reduce'

The function 'β*' is now called 'lazy-allreduce' and has slightly different
semantics.

The 'τ' macro is now called 'transform'.  It also has a new syntax that
avoids two pairs of parentheses.  Instead of (τ (i j) (j (1+ i))), you can
now write (transform i j to j (1+ i)).  As a bonus, the new macro also
supports input constraints that are not literal integers.  The syntax for
introduce such a constraint is (transform (variable constraint) ... ).

Sorry for breaking your code.  I promise this was the last major
incompatible change.