INCOMPATIBLE CHANGE: Change broadcast semantics, improve reshaping.
When broadcasting lazy arrays to shapes with higher rank, axes are added at
the end instead of at the beginning.  This is different than what NumPy
does, but fits better into the leading axes paradigm of Petalisp.

The lazy-reshape function has seen a major overhaul.  It is now possible to
explicitly set the number of relevant axes with an integer modifier.  All
axes above that number are silently passed along.  This makes it possible
to write that supports batch computing by means of extra trailing axes.
Also, it is now possible to use functions as modifiers, in which case they
are expected to take a shape consisting of all the relevant axes, and
return any number of new modifiers as multiple values.  We call such
functions reshapers. The standard library already provides three reshapers:
collapsing-reshaper, peeling-reshaper, and permuting reshaper.

Another thing that has changed about the lazy reshape operator is that it
is more strict about selecting or broadcasting values.  It used to align
ranges with the same size but different elements automatically, but will
now signal an error.  Consider using the collapsing reshaper to handle such
cases.