Group store instructions into stencils, too.
Although having stencil patterns while storing is much less frequent, it will
still improve the performance in some cases.  In addition, this change makes
kernels symmetric in terms of how their load and store instructions are
represented.

This commit also defines a partitioning visualizer, which I used for debugging
all the problems into which I stumbled when writing this commit.