INCOMPATIBLE CHANGE: Simplify the semantics of lazy-reshape.
This commit changes the semantics of lazy-reshape so that it won't allow
lexicographical reordering of elements to a new shape anymore.  The
functionality is of course still available, but via the newly introduced
function lazy-rearrange.  Furthermore, lazy-reshape stops supporting integer
modifiers.

The upside of these changes is that the number of axes being worked on is
always clear from the modifier: For transformation modifier, the number of axes
being actively manipulated is the input rank of that transformation.  For a
shape modifier, the number of axes being actively manipulated is the rank of
that shape.  Because of this, the clumsy notion of 'relevant shape' could be
dropped entirely from the description of lazy-reshape.

A follow-up improvement is that lazy-reshape will now also support modifiers
that don't match the rank of the lazy array being worked on.  In those cases,
it applies the usual broadcasting rules.  The result of these changes is that
it is now much easier to write code that supports implicit batches.