Make lazy-reduce more powerful.  Eliminate lazy-multireduce.
These days, one golden rule of programming in Petalisp is to never infer the
rank of an operation from the rank of a supplied array.  This rule is crucial
to allow implicit batches.  The function lazy-multireduce broke this rule, and
has therefore be removed.  To make up for it, the function lazy-reduce has been
upgraded to allow any fixed number of successive reductions instead of just one.