INCOMPATIBLE CHANGE: Simplify reshaper naming scheme and semantics.
The individual changes introduced by this commit are:

- The collapsing-reshaper is now called deflater.  More general, the
terminology has been changed throughout the code base so that the act of moving
a lazy array to have a start of zero and a step of one is called 'deflating'
instead of 'collapsing'.

- The permuting-reshaper is gone, because it produces no extra value over the
plain use of make-transformation.

- The peeling-reshaper has been renamed to peeler, and its semantics have been
changed so that it receives amount designators instead of multiple keyword
arguments.  This change encourages the good practice of being explicit about
the number of axes being worked on.