Generate Lisp code instead of C++ code by default
Modern C++ compilers are much slower than SBCL, which makes for a poor user
experience when working at the REPL.  The plan for the future is to compile
only kernels with large iteration spaces to C++ and generate Lisp code
otherwise.  Until the necessary machinery for adding such kernel metadata is in
place, use the Lisp compiler only.

Users looking for that extra boost in performance only have to change the
native-backend's mixin to lisp-plus-cpp-compiler-mixin.